move server stream mock outside of requestid for reuse
